refactor(file-naming): standardize file names across all components, part 2
