+++ frontend/app/[locale]/(routes)/stories/_components/TestSection.tsx
@@ -1,5 +1,5 @@
 import TestChoice from './TestChoice';
-import TestDescription from './testDescription';
+import TestDescription from './TestDescription';
 import TestInstruction from './TestInstruction';
 
 const TestSection = () => {
